ON-16269: kernel implementation of shrub client
Abstract the socket operations used by shrub_client.c into a new
module shrub_socket.c, and provide an implementation for use in
the kernel.

+++ src/lib/ciul/shrub_client.h
@@ -13,9 +13,9 @@
 /* Structure for managing a client instance */
 struct ef_shrub_client
 {
-  int      socket;
-  int      files[EF_SHRUB_FD_COUNT];
-  uint64_t mappings[EF_SHRUB_FD_COUNT + 1];
+  uintptr_t socket;
+  uintptr_t files[EF_SHRUB_FD_COUNT];
+  uint64_t  mappings[EF_SHRUB_FD_COUNT + 1];
 };
 
 /* Request shared rxq token from shrub server
